+++ effigies/src/main/java/org/identigon/effigies/ScaffoldCommand.java
@@ -72,7 +72,7 @@ class ScaffoldCommand {
             out.println("Scaffold written to " + file);
             return 0;
         } catch (Exception e) {
-            err.println("Error: " + e);
+            err.println("Error: " + CliErrors.causeChain(e));
             return 1;
         }
     }